requirements.txt
django==4.1.3
djangorestframework==3.14.0
drf-nested-routers==0.93.4
psycopg2--binary==2.9.5
python-dotenv==0.19.2

